Increase max _id size from 17 to 24

+++ app/koru/user-account/main-server-test.js
@@ -186,7 +186,7 @@ isServer && define(function (require, exports, module) {
       },
 
       "test expired token": function () {
-        assert.equals(userAccount.model.$fields.resetTokenExpire, {type: 'integer'});
+        assert.equals(userAccount.model.$fields.resetTokenExpire, {type: 'bigint'});
 
         v.lu.resetToken = 'secretToken';
         v.lu.resetTokenExpire = Date.now() -5;
